+++ cmd/search.go
@@ -11,9 +11,10 @@ import (
 )
 
 var searchCmd = &cobra.Command{
-	Use:               "search [<search-term> ...]",
-	Short:             "Search full text of TMs in catalog using bleve search engine",
-	Long:              `Search full text of TMs in catalog using bleve search engine`,
+	Use:   "search [<search-term> ...]",
+	Short: "Search full text of TMs in catalog using bleve search engine",
+	Long: `Search full text of TMs in catalog using bleve search engine. For each repository to be searched,
+a local search index has to be created once using 'create-si' command.`,
 	Args:              cobra.MinimumNArgs(1),
 	Run:               executeSearch,
 	ValidArgsFunction: completion.CompleteTMNames,