+++ internal/app/http/mapping.go
@@ -113,7 +113,7 @@ func (m *Mapper) GetInventoryEntryVersion(version model.FoundVersion) server.Inv
 		invVersion.Attachments = &atts
 	}
 
-	if version.SearchScore != 0.0 {
+	if version.SearchScore != 0 {
 		invVersion.Score = &version.SearchScore
 	}
 	if len(version.MatchLocations) > 0 {